+++ changelog.md
@@ -1,3 +1,9 @@
+
+v1.1.1
+======
+Added: Setting to input a custom demo site URL
+Added: Support for site migration using dynamic demo site URL
+
 v1.1.0
 ======
 Added: Auto migration using demo site URL